Finish working on lifecycle, and delete image:
  - [x] Add DeleteContainer()
  - [x] try http package instead of 3rd party http client: http://golang.org/pkg/encoding/json/#example_Unmarshal
        => need to replace more, but good for now
  - [x] showImage use image:tag param
  - [x] Check if container already exists, delete if so then create image

+++ lib/types/types.go
@@ -4,7 +4,7 @@ import Logger "../logging"
 
 type ApiContainer struct {
   Command string
-  Created int
+  Created int64
   Id string
   Image string
   Names []string
